Coding style cleanup (phpcs). with some other cleanup
Admin layout must not be set using session variable.
thus, the methods have been removed in Users controller
and AppController

Admin is now split and using prefix, however the routing
is not matched accordingly yet and layout management is
not either

action names have been camelCased everywhere. this should
be matched with DashedRoute